+++ book-service/src/main/java/com/example/demo/models/entity/Book.java
@@ -1,4 +1,4 @@
-package com.example.models.entity;
+package com.example.demo.models.entity;
 
 import java.util.Objects;
 import java.util.Set;
