+++ book-service/src/main/java/com/example/demo/models/repository/BookRepositoryInMemory.java
@@ -1,13 +1,12 @@
-package com.example.models.repository;
+package com.example.demo.models.repository;
 
-import com.example.models.entity.Book;
+import com.example.demo.models.entity.Book;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.stream.Collectors;
